# packages/server/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Kernl Server API",
    "description": "HTTP API for @kernl-sdk/server"
  },
  "paths": {
...
    "/threads": {
      "get": {
        "operationId": "threads.list",
        "parameters": [
          {
            "in": "query",
            "name": "agentId",
            "schema": {
              "type": "string"
            },
            "description": "Filter by agent ID"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Max results"
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            },
            "description": "Pagination cursor"
          }
        ],
        "summary": "List threads",
        "description": "Get a list of all kernl threads, sorted by most recently updated.",
        "responses": {
          "200": {
            "description": "Paginated threads",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedThreads"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "threads.create",
        "parameters": [],
        "summary": "Create thread",
        "description": "Create a new kernl thread for interacting with AI assistants and managing conversations.",
        "responses": {
          "200": {
            "description": "Created thread",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Thread"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestError"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadCreate"
              }
            }
          }
        }
      }
    },
    "/threads/{tid}": {
      "get": {
        "operationId": "threads.get",
        "parameters": [
          {
            "in": "path",
            "name": "tid",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "summary": "Get thread",
        "description": "Retrieve thread metadata.",
        "responses": {
          "200": {
            "description": "Thread",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Thread"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "threads.delete",
        "parameters": [
          {
            "in": "path",
            "name": "tid",
            "schema": {
              "type": "string",
              "pattern": "^ses.*"
            },
            "required": true
          }
        ],
        "summary": "Delete thread",
        "description": "Delete a thread and permanently remove all associated data, including messages and history.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "threads.update",
        "parameters": [
          {
            "in": "path",
            "name": "tid",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "summary": "Update thread",
        "description": "Update properties of an existing thread, such as title or other metadata.",
        "responses": {
          "200": {
            "description": "Updated thread",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Thread"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadUpdate"
              }
            }
          }
        }
      }
    },
...
    "/threads/{tid}/messages": {
      "get": {
        "operationId": "threads.messages.list",
        "x-note": "AI SDK adapter endpoint. Returns UIMessages. The future kernl-native equivalent will be GET /threads/{tid}/events.",
        "parameters": [
          {
            "in": "path",
            "name": "tid",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "summary": "Get thread messages",
        "description": "Retrieve messages for a thread as UIMessages.",
        "responses": {
          "200": {
            "description": "Thread messages",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AISDKMessageList"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      }
    },
    "/threads/{tid}/stream": {
      "post": {
        "operationId": "threads.stream",
        "x-note": "AI SDK adapter endpoint. Accepts UIMessage, streams UIMessage responses. The future kernl-native equivalent will be POST /threads/{tid}/events.",
        "summary": "Stream thread execution",
        "description": "Send a message to a thread and stream the agent response. The thread must exist and determines which agent handles the request.",
        "parameters": [
          {
            "in": "path",
            "name": "tid",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Thread ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AISDKStreamRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SSE stream of UI messages",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "Server-Sent Events stream"
                }
              }
            }
          },
          "404": {
            "description": "Thread not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Thread": {
        "type": "object",
        "properties": {
          "tid": {
            "type": "string",
            "description": "Unique thread identifier"
          },
          "namespace": {
            "type": "string",
            "description": "Namespace for multi-tenancy"
          },
          "title": {
            "type": "string",
            "nullable": true,
            "description": "Thread title"
          },
          "agentId": {
            "type": "string",
            "description": "ID of the owning agent"
          },
          "model": {
            "$ref": "#/components/schemas/ModelInfo"
          },
          "context": {
            "type": "object",
            "additionalProperties": true,
            "description": "User-defined context"
          },
          "parentTaskId": {
            "type": "string",
            "nullable": true,
            "description": "Parent task ID if spawned"
          },
          "state": {
            "$ref": "#/components/schemas/ThreadState"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "tid",
          "namespace",
          "agentId",
          "model",
          "state",
          "createdAt",
          "updatedAt"
        ]
      },
      "ThreadState": {
        "type": "string",
        "enum": ["idle", "running", "stopped", "error"],
        "description": "Thread lifecycle state"
      },
      "ThreadCreate": {
        "type": "object",
        "properties": {
          "tid": {
            "type": "string",
            "description": "Optional thread ID (auto-generated if omitted)"
          },
          "agentId": {
            "type": "string",
            "description": "ID of the agent"
          },
          "title": {
            "type": "string",
            "description": "Optional title"
          },
          "context": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": ["agentId"]
      },
      "ThreadUpdate": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "context": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "AISDKMessageList": {
        "type": "object",
        "x-note": "AI SDK adapter type. Will be replaced by kernl-native event types in the future.",
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UIMessage"
            },
            "description": "Messages as UIMessages"
          }
        },
        "required": ["messages"]
      },
      "UIMessage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": ["user", "assistant"]
          },
          "parts": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "Message parts (text, tool calls, etc.)"
          }
        },
        "required": ["id", "role", "parts"]
      },
...
      "ModelInfo": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "description": "Provider ID"
          },
          "modelId": {
            "type": "string",
            "description": "Model ID"
          }
        },
        "required": ["provider", "modelId"]
      },
...
      "PaginatedThreads": {
        "type": "object",
        "properties": {
          "threads": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Thread"
            }
          },
          "next": {
            "type": "string",
            "nullable": true,
            "description": "Cursor for next page"
          }
        },
        "required": ["threads"]
      },
...
      "Success": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "const": true
          }
        },
        "required": ["success"]
      },
      "AISDKStreamRequest": {
        "type": "object",
        "x-note": "AI SDK adapter type. Will be replaced by kernl-native event types in the future.",
        "description": "Request body for AI SDK streaming endpoints.",
        "properties": {
          "message": {
            "$ref": "#/components/schemas/UIMessage",
            "description": "UIMessage to send"
          }
        },
        "required": ["message"]
      }
    }
  }
}